Attributes and Ignore Files
I have uploaded the GitAttributes and GitIgnore files as they, for whatever reason, were not playing nice with uploading through the GitHub website.

They should now be in the repo, and the repo should now be ready for adding the Unity Project to it, assuming we are using the Unity Game Engine.
If we are not using Unity, we will need to find a new .gitIgnore file for that engine, but there should be one online