Base States & Test Enemy
I have created a handful of Base States, which are states that should be able to be shared by every entity in the game, though not all will necessarily use them.

Additionally, I have created a Test Enemy that utilises these states and can transition between them based on conditions (Currently using booleans within the TestEnemy script or that are being read from the states themselves).
I have done a new way of implementing the states into a brain by declaring them as variables of the brain. This allows for editing of the state's parameters during runtime, however it may restrict the use of the readonly attribute. This shouldn't affect us much so long as we keep it in mind as we progress forwards, and we may find an alternative once we look.

Furthermore, I have created a Position SO and a PositionWriter component which will allow for easy access of a Transform's position (Such as the Player) without requiring a reference to the transform itself.

Current States:
- Unaware (Sub State Machine): No Additional Logic
- Combat (Sub State Machine): No Additional Logic
- Dead (State): Destroys the '_objectToDestroy' GameObject after '_destroyDelay' seconds
- Idle (State): Sets a completion time in OnEnter and has a getter to return whether the duration has elapsed.
- Patrol (State): No Additional Logic
- Chase (State): No Additional Logic. We will need to figure out how to pass the target to this state.