Initial Unit Movement
I have implemented a basic unit movement for the A* Pathfinding System, however I am going to improve this to make it easier for implementing new versions of pathfinding.

We can currently have a unit navigate to a target position at a desired speed, avoiding obstacles based on a path generated from the A* Pathfinding solution, with the waypoints simplified slightly based on angle requirements.