Path Request Manager Decoupling
I have changed how the PathRequestManager handles its reference to the Pathfinding script, now allowing it to use other types of Pathfinding so long as those scripts inherit from the IPathfinder interface. This will allow easier integration of new pathfinding systems should we choose to use them.

Additionally, I have removed the requirement of a reference from the Pathfinding script to the PathRequestManager by having it instead use a passed callback to output the resulting path.

Finally, I have also changed the way that the path is returned, using a struct which stores the array of Vector2 waypoints as well as the boolean for whether the path is valid or not.