New Input System
I have transitioned the project to the New Input System, putting an example implementation of how it'd work in the PlayerMovement script while keeping the original implementation commented inside.

Additionally, I made it so that the canvas scales with screen size. Sorry, but it was annoying me and I'm bad at ignoring things