Player Attack Start
I have created a PlayerAttack script which receives input from the PlayerInput class and supports functionality for combos with the primary attack (E.g. Primary 1 > Primary 2 > Primary 3 > Primary 1)

The Attack class is currently just a standard script with a 'name' variable so that it shows up in the inspector, but I'm waiting for input from others on how we should store the attacks before I continue.
The Attack class itself will be a base from which others inherit (Potentially an abstract class).