FaceDirection Fix & Spelling Correction
I have fixed an issue in the FaceDirection script which was causing the player to not rotate to face the moues when the mouse wasn't moving but the player was.

Additionally, I fixed a spelling mistake in the PlayerControls InputActionAsset ('FaceDiraction' changed to 'FaceDirection'). I also changed the dash on the gamepad to the Right Shoulder Button (R1 on DS5) to match what was laid out in the GDD.

Finally, I resolved the issue that was preventing the switching of control schemes between MnK and Gamepad (I set 'JoiningEnabledByDefault' to false on the PlayerInputManager component)