Avoidance vs Keep-Distance Work
I have worked on the Avoidance and Keep Distance behaviours to try and get them working together, though I haven't completed this goal as of this commit.

I have also changed how the Gizmos are drawn, giving each behaviour a DrawGizmos function that is called from the State they are used in.