CBS Work
I have done a little work on the CBS Avoidance and Keep Steering Behaviours. Not much progress was made, as we tried adding a interest to the avoidance behaviour facing diagonally away from the obstacles, however this caused the entity to bounce in and out of range.

The changes have been commented out, but haven't been removed