Quick Upload
Idea: For interest, take the highest value. If there is a tie, take the one with the lowest dot compared to the targetDir.
Idea 2: Don't sum the direction directly opposite the largestInterestDir.

Idea: Pass in a Rigidbody2D in place of the position so that we can get information such as position, rotation, & velocity. This will allow us to have the strafe behaviour favour moving a similar direction to our velocity, and when this changes (E.g. Repelled by a wall) so too will our movement direction.
Idea 2: Pass in a struct for the targetInformation, with the struct containing a Vector2 position & Vector2 facingDirection.

Idea: Return a struct in place of the floats, with the struct containing a float-bool pair. The float represents the strength of the interest, and the bool represents whether the direction should be normalized.