Decoupled Health UI & Merge Conflict
I have created a new Health UI that is decoupled from the player, allowing the player to function without its existance. This PlayerHealthUI also uses hearts which can have half values.

Note: The reason I have called it 'PlayerHealthUI' as opposed to simply 'HealthUI' is due to the fact that only the player will use hearts for their health. If we eventually add bosses, we discussed about them having a 'Bar' as opposed to hearts.

Note 2: I have kept the old 'PlayerHealth' script on the player so that we can eventually copy its respawning functionality if we so wish.