UI Manager & More Health Bars
I have created a UI Manager which can currently create a Health Bar for use by a boss enemy.

Additionally, I have created a ProgressBar script which can be used for Health Bars, Stamina Bars, Ability Cooldowns, etc; along with a variation on the HealthBarUI which assigns values to a ProgressBar as opposed to the PlayerHealthUI which does hearts.

Note: I haven't made the PlayerHealthUI automatically subscribe itself to the player's HealthComponent yet, however we should think about whether we actually want this to happen or if we should just leave it as a HeartHealthUI and manually assign (Reduces coupling)