Patrol Fix & AoE Speed
We fixed a bug with the patrol state where it would return an error if there were no set patrol points. With the designers' current implementation of the enemies (Using a TestEnemy that has a patrol state, even for enemies that don't patrol), this was an issue, however for the future it shouldn't be.

Furthermore, we changed the AoE Attack to use a speed rather than a fixed time for its detonation. We also made it so that this cannot pass through walls.
In the next change we could make it so that the AoE Indicator (And subsequent explosion) is essentially a projectile, which would also allow for it to be deflected.