Player Weapon UI
I have created a script to manage the UI for the Player's Weapons. It subscribes to the PlayerAttacks.cs script when it is enabled (Unsubscribing when disabled), though currently the events for Recovery Time & Use Recharge Time are called every frame, even when they are not changing.

The current functions of the PlayerWeaponUI are:
- Shows Recovery for both Primary and Secondary Weapons
- Shows Recharge Time for both Primary and Secondary Weapons
- Updates to show the sprite of the assigned Primary and Secondary Weapon

Furthermore, I have added an option for weapons to only start recharging their uses if they have ran out, which would be used for things like the Revolver.