Explosive Projectiles
I have changed the AoE Attacks to use an explosive projectile, which has options for explosive delay, exploding on collision, reducing the radius if it explodes early, colliding with the environment (Including the ability to reflect off of walls), and displaying a viewer for the max radius with a centre increasing to that size as the delay nears 0.

I have also added more features to the standard projectiles, including giving them the capability to move towards a set position & home towards a target transform (Both of which are compatible with Reflection).

Furthermore, the AoERadiusViewer can be set to have different curves used for the growth of the radius, with the default being a linear growth.