Dash Change & Inspector Prettying
We have change how we are handling the dash from changing the player's speed to setting the velocity of the player's rigidbody. This has allowed us to have the player dash even when stationary.

Furthermore, I have exposed some additional variables to the inspector and renamed some others, providing more clarity through their names and giving the designers more control over aspects of the dash such as how soon after dashing the stamina starts regenerating (That's it for the newly exposed variables).

In addition to these changes, I have added [Header] and [Space] attributes to make the variables more organised in the inspector.