Linking Animations to Controller (Player)
I have linked the player to be able to play animations for attacks, idle, running, hurt, & dead.

Animations play at a fixed framerate, though we can later have it that you can change the framerate in runtime. Animations can be set to loop, set as temporary, set to be able to be overriden by anything or to override anything, and (For temporary anims) can be given a condition in the form of a Func<bool> to delay their transition back to the previous animation (E.g. Attack Animations wait until their attack duration has elapsed before allowing reversion, even if the animation ends before)