Assigned Layers & Police Melee
Due to the creation of the new 'Entity' layer, some things (Such as attacks & vision) weren't correctly assigned. Everything should now be working with these new layers.

Additionally, I fixed a bug in the HealthComponent where an entity would not take damage from two sources if the damage occurred during the same frame, even if their IFrameDuration was set to 0.

Finally, I created the Police Melee, which is essentially a tougher version of the Yakuza Melee but is on the 'Police' faction. For this, I removed the 'Test Faction' faction and added in the 'Police' faction along with a currently unused 'RivalYakuza' faction.