Weapon Animations
I started working on the weapon animations, and they currently function for both the enemies and the player. They aren’t applied to the prefabs yet, but they are functioning for the Katana and Tanto (Player & Yakuza Melee weapons).

I’ve also linked the animations to all the existing Entity Brains. Currently, we are using both events and a reference, however I am likely to keep it as only a reference.