Title Screen Start
I have started work on the Title Screen scene, adding in the Main Menu, Tutorial Scene Selection Modal, Options Screen Tabs (Not the functionality of the settings, just tab switching).

I have also added a Persistent Scene to house the GameManager and Loading Screen. (The GameManager handles the loading of scenes so that we can have a Loading Screen in the game, and it automatically loads the TitleScreen scene when initialised)