Loading Screen Update
I have updated the loading screen so that it displays the current loading percentage and then 'Press Any Key to Continue' once the scenes are ready to play (With an additional 0.5 realtime-second delay to ensure objects are ready).

I've also added a Camera to the Persistent Scene so that the loading screen displays properly. (Note: I spent a while messing about with cameras not realising that I hadn't yet uploaded this, so there may be some issues)