Persistent Camera & Changing Active Scenes
I’ve moved the camera into the Persistent Scene and made it so that the GameManager has a reference to it which other scripts can access.

Additionally, I have made it so that the GameManager can choose to set a scene that it is loading to the active scene, given its build index.