UI Pivots & Bloom Toggle Functionality
I have made the UI in the PlayerTestScene have pivots that allow for easy scaling through changing the scale of the root GameObject. This can allow for us to implement a UI scale option in the future.

I have also added functionality to the Bloom toggle in the Graphics settings, however there are definitely better ways to do it that what I have. Also, we should look at making an Exception Formatter for our Singletons.