Tanto, Baton, Shotgun, & Player Outline
I've added in Sprites & Animations for the Tanto, Baton, & Shotgun, along with a Weapon and Attack for the Shotgun.

Additionally, I have added the Player Outline again, using a new Circle sprite which allows for room at its edges for the Outline. Furthermore, I fixed an issue with the Options Menu which led to the Alpha of the colour always being set to 0.