Added Player Scene & Improved Scene Loading
I've added a new Scene to store the Player, Player Canvas, and Player Manager.

Additionally, I have improved the SceneTransitions, allowing you to specify a spawn position for the player. Furthermore, I have added a new type of transition which is dedicated to transitioning FROM the Title Screen, and can automate things like adding the Player Scene and removing the Title Screen.