Dialogue Start & Clone Brain
I have started work on the Dialogue, adding in Scriptable Objects to represent the conversations, a DialogueManager to process said ScriptableObjects (Complete with skipping & continuing), a DialogueUI to process the UI elements, UI to show the dialogue, and the ability to pause things such as Enemy AI and Player Input while the UI is active.

Additionally, I have also changed the Clones over to use a separate brain from the enemies. This will improve performance marginally, due to slightly less overhead when checking the states, but more importantly it removes the Attacking state from them, as their lack of attacks was causing errors.
If we later wish for clones to be able to attack, we can add this state into their brain and add checks to ensure that no errors are thrown for Weapons without Attacks.