Scoreboard Scene Fix
I've fixed issues with the Scoreboard scene that arose from it using Transforms as opposed to RectTransforms for UI Elements, along with the canvas not being properly set up for scaling with varying screen sizes and aspect ratios, text not having TextElement scripts attatched, and that many of the UI Elements were using different scales as opposed to changing the widths and heights of the RectTransform.

I also cleaned up the hierarchy a little, organising the scores by type and then grouping them all together under a parent which also acts as their layout group.