Weapon Fixes & Enemy Rotation Change
I've fixed some bugs with the Attacks and Weapons that was causing them to not rotate in the direction of their rotation pivot.

Additionally, I have changed the enemies so that they only rotate their RotationPivots, as opposed to their entire Transform. For this, I've changed the EnemySenses and EnemyMovement scripts.