Attack Interruption & Target Missing Error Fix
I've added in the ability for attacks to be interrupted when the entity exits the Attacking state (E.g. When they are stunned). I have also changed the Attacking, ComplexAttacking, and HealthBasedAttacking states to only allow exits when they are not in the process of attacking, and have made the Stunned and Dead transitions override this.

I have also fixed a potential error in the StandardEnemy and ComplexAttackingEnemy brains where the EntitySenses could occasionally attempt to find its target between the FSM checking for transitions and the state logic being run. In the event that it lost its target, an error would be thrown when we try to retrieve the target's position, as the transition to the Unaware state wouldn't occur till the start of the NEXT frame.