Knockback & Kickback + AoEAttack CanHitSelf Fix
I've added the ability for ALL attacks to apply knockback on a hit, along with allowing attacks to apply a kickback/kickforward force to their attacker.

Furthermore, I fixed the issue with the AoEAttack which was preventing it from ignoring the IgnoreTransform when executing the explosion.

I've also fixed (Or tried to fix) the 'Shotgun Sprite' prefab