Player Animations & EntityAnimator change
I've implemented the player's idle animations, movement animations for East & West, and dashing animations for East and West.

I've also changed the EntityAnimator script so that the player's input overrides the value of the RotationPivot, allowing them to face in the direction they are moving as opposed to the direction they are facing. We can revert this if wished.