Weapons & Player Movement Animation Update
I've updated the weapon and movement animations for the player, along with fixing a few bugs.

Weapons can now prevent the player from moving for the duration of their attack, and their animations now use Animator.CrossFade as opposed to triggers, integers, and transitions. We still need to assign transitions back to Idle, however, but it is a good enough compromise.

We changed the Katana to be in more of a guard pose when idle, and re-animated its attacks to follow accordingly. Additionally, we changed the Katana Slash 3 attack's Recovery Time (0.55s > 0.33s) and Cooldown Time (0s > 0.55s) to allow them slight movement between the end of the final slash and the start of the next combo.

We also changed the PlayerFaceDirection so that if we are using a gamepad but are not pushing the Right Stick, we will attempt to use the movement input to determine direction.
Furthermore, the player no longer rotates towards their direction of movement by default, and instead always faces towards the mouse/gamepad right stick direction, unless the situation described above is in effect.