Respawning Change
I've changed the Player Respawning to now function via the GameManager reloading scenes. There is an option to only reload the player's scene, but by default we reload every scene that isn't the Persistent Scene.

I've also made it so that the CheckForAbility script doesn't cache the Player's GO, and that the PlayerFollowCameraInitialiser doesn't destroy itself and instead of initialising itself on Awake/Update, it instead initialises the camera when the OnResumeLogic event is triggered by the GameManager.

Additionally, I've changed the Tattoo Selection UI so that it hides the camera view (Preventing you from seeing it as it transitions over to the starting position)